backend/requirements.txt
asgiref==3.11.0
bcrypt==5.0.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.15.2
google-api-core==2.28.1

google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.4
stripe==11.4.0
urllib3==2.6.1
whitenoise==6.11.0
boto3==1.35.0
django-storages==1.14.2